Made python reqs into an action
