+++ .builder/actions/manylinux1-ci.py
@@ -1,6 +1,7 @@
 
 import Builder
 import os, sys
+from aws_crt_python import InstallPythonReqs
 
 pythons = (
     #'cp34-cp34m',
@@ -23,7 +24,7 @@ class ManyLinux1Package(Builder.Action):
         for version in pythons:
             python = python_path(version)
             actions = [
-                [python, '-m', 'pip', 'install', '--upgrade', 'autopep8', 'boto3'],
+                InstallPythonReqs(python=python),
                 [python, '-m', 'pip', 'install', '.', '--install-option=--verbose', '--install-option=sdist', '--install-option=bdist_wheel'],
                 ['auditwheel', 'repair', '--plat', 'manylinux1_x86_64', 'dist/awscrt-*{}-linux_x86_64.whl'.format(python)],
             ]
@@ -47,9 +48,7 @@ class ManyLinux1CI(Builder.Action):
             python = python_path(version)
             
             actions = [
-                [python, '-m', 'pip', 'install', '--upgrade',
-                    "--trusted-host", "pypi.org", "--trusted-host", "files.pythonhosted.org",
-                    "pip", "setuptools", "boto3"],
+                InstallPythonReqs(python=python, deps=['boto3'], trust_hosts=True),
                 [python, '-m', 'pip', 'install', '.',
                     '--install-option=--verbose', '--install-option=build_ext', '--install-option=--include-dirs{openssl_include}',
                     '--install-option=--library-dirs{openssl_lib}'],
